# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: MyCRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities with optional filtering
      parameters:
        - name: contactId
          in: query
          schema:
            type: string
        - name: dealId
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated activity list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedActivities"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Manually log an activity (note, call, meeting)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityWithRelations"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Campaigns ──────────────────────────────────────────────────────────────
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List email campaigns
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CampaignStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated campaign list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCampaigns"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create an email campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignWithStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign with stats
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignWithStats"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update or send a campaign
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignInput"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignWithStats"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /campaigns/{id}/send:
    post:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Send a campaign to its recipients
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendCampaignInput"
      responses:
        "200":
          description: Sending result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignWithStats"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Enums ──────────────────────────────────────────────────────────────
    ContactStatus:
      type: string
      enum: [LEAD, PROSPECT, CUSTOMER, CHURNED, UNQUALIFIED]
...
    ActivityType:
      type: string
      enum:
        [
          NOTE,
          CALL,
          EMAIL_SENT,
          EMAIL_OPENED,
          EMAIL_CLICKED,
          DEAL_CREATED,
          DEAL_MOVED,
          DEAL_WON,
          DEAL_LOST,
          TASK_CREATED,
          TASK_COMPLETED,
          CONTACT_CREATED,
          MEETING,
        ]
    CampaignStatus:
      type: string
      enum: [DRAFT, SCHEDULED, SENDING, SENT, CANCELLED]
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required: [id]
...
    ContactSummary:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ContactStatus"
      required: [id, firstName, lastName, status]
...
    ActivityWithRelations:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/ActivityType"
        title:
          type: string
        description:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        userId:
          type: string
          nullable: true
        user:
          $ref: "#/components/schemas/UserSummary"
          nullable: true
        contactId:
          type: string
          nullable: true
        contact:
          $ref: "#/components/schemas/ContactSummary"
          nullable: true
        dealId:
          type: string
          nullable: true
        deal:
          type: object
          nullable: true
          properties:
            id:
              type: string
            title:
              type: string
          required: [id, title]
        createdAt:
          type: string
          format: date-time
      required: [id, type, title, createdAt]
    CampaignStats:
      type: object
      properties:
        total:
          type: integer
        sent:
          type: integer
        opened:
          type: integer
        clicked:
          type: integer
        openRate:
          type: number
        clickRate:
          type: number
      required: [total, sent, opened, clicked, openRate, clickRate]
    CampaignWithStats:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        subject:
          type: string
        fromName:
          type: string
        fromEmail:
          type: string
        htmlContent:
          type: string
        textContent:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/CampaignStatus"
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        stats:
          $ref: "#/components/schemas/CampaignStats"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        [id, name, subject, fromName, fromEmail, htmlContent, status, stats, createdAt, updatedAt]
...
    PaginatedActivities:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActivityWithRelations"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean
      required: [data, total, page, pageSize, hasMore]
    PaginatedCampaigns:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CampaignWithStats"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean
      required: [data, total, page, pageSize, hasMore]

    # ── Input schemas ──────────────────────────────────────────────────────
...
    CreateActivityInput:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/ActivityType"
        title:
          type: string
        description:
          type: string
        contactId:
          type: string
        dealId:
          type: string
        companyId:
          type: string
      required: [type, title]
    CreateCampaignInput:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        fromName:
          type: string
        fromEmail:
          type: string
        htmlContent:
          type: string
        textContent:
          type: string
        recipientContactIds:
          type: array
          items:
            type: string
      required: [name, subject, fromName, fromEmail, htmlContent]
    UpdateCampaignInput:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        fromName:
          type: string
        fromEmail:
          type: string
        htmlContent:
          type: string
        textContent:
          type: string
    SendCampaignInput:
      type: object
      properties:
        recipientContactIds:
          type: array
          items:
            type: string