# lib/api-zod/src/generated/types/activityType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyCRM API specification
 * OpenAPI spec version: 0.2.0
 */

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  NOTE: 'NOTE',
  CALL: 'CALL',
  EMAIL_SENT: 'EMAIL_SENT',
  EMAIL_OPENED: 'EMAIL_OPENED',
  EMAIL_CLICKED: 'EMAIL_CLICKED',
  DEAL_CREATED: 'DEAL_CREATED',
  DEAL_MOVED: 'DEAL_MOVED',
  DEAL_WON: 'DEAL_WON',
  DEAL_LOST: 'DEAL_LOST',
  TASK_CREATED: 'TASK_CREATED',
  TASK_COMPLETED: 'TASK_COMPLETED',
  CONTACT_CREATED: 'CONTACT_CREATED',
  MEETING: 'MEETING',
} as const;